Storage module rewrite
Rewrote storage module and changed everything that uses it. localStorage is hard limited to 5MB and the map data needs to be saved into storage and I believe will be over 4MB. I decided that this wasn't a reasonable situation, so I rewrote the storage module to instead use indexedDB to store the data.

No data migration is included.