Serve files via node
Working towards having a server component that can continue running even when the client isn't so that automated play can continue 24/7. In pursuit of this, we will now be serving the website via nodejs. Planned next steps are to move API requests and primary storage from the client to the node backend. Then add map downloading functionality and have it automatically download new map data when the server refreshes.