Enable loading of star chart data
Both adds functionality for the server to load and cache all the star chart data and also adds the necessary functionality for the client to display and update star chart loading progress. When there is a ST server reset the local server will clear out all the stale chart data and the client will disable the star chart page. If the agent server reset behavior is configured to automatically recreate agents and it is successful in doing this for the selected user, the server will automatically start re-loading the star chart data after this happens.

Additionally creates an interface for components to mark themselves as up-to-date with the current server reset for components that, unlike the agent component, do not want to store that information in a per-record fashion.

Eventually we will probably need a way of loading a system's waypoints as a one-off, so that functionality will probably need to be refactored out of `start_chart_load` so that we can load that data as a one-off when loading all waypoint data isn't yet complete.