Refactor to prep for universe rendering
A lot of the star chart data was written specifically for rendering waypoints without much thought as to how the universe would eventually be rendered. Now we have to reconsider that design, doing things like pulling rendering data out of the `#current_system.waypoints` entries and putting it in a place that can be common to waypoints any systems.

Also adds tiny bits of functionality that will be needed for universe rendering such as declaring `#local_universe` and `#universe_bounds` and loading `#universe_bounds` near where we will start universe rendering.